+++ packages/motion/src/error.js
@@ -6,9 +6,9 @@ export const MESSAGES = {
   ALREADY_EXECUTING: 'Another instance of motion is already running that directory in the same process',
   ALREADY_MOTION_APP: 'Directory is already a motion app'
 }
+
 // Note: This is generating automatically from MESSAGES
 export const ERROR_CODE = {}
-
 for (const name of Object.keys(MESSAGES)) {
   ERROR_CODE[name] = name
 }