+++ packages/motion/src/index.js
@@ -1,32 +1,30 @@
 /* @flow */
 
-import invariant from 'assert'
 import Path from 'path'
-import { exists, copy, mkdir, realpath, writeFile } from './fs'
+import invariant from 'assert'
 import { CompositeDisposable, Disposable, Emitter } from 'sb-event-kit'
-import State from './state'
+import * as FS from './fs'
 import CLI from './cli'
+import Config from './config'
 import { MotionError, ERROR_CODE } from './error'
-import { fillConfig, getPundleInstance } from './helpers'
-import type { Motion$Config } from './types'
+import { getPundleInstance } from './helpers'
 
 class Motion {
   cli: CLI;
-  state: State;
-  config: Motion$Config;
+  config: Config;
   emitter: Emitter;
   watching: boolean;
+  projectPath: string;
   subscriptions: CompositeDisposable;
 
-  constructor(state: State, config: Motion$Config) {
-    invariant(state instanceof State && typeof config === 'object',
-      'Use Motion.create instead of constructor')
+  constructor(projectPath: string, config: Config) {
+    invariant(config instanceof Config, 'Use Motion.create instead of constructor')
 
-    this.cli = new CLI(state, config)
-    this.state = state
+    this.cli = new CLI(config)
     this.config = config
     this.emitter = new Emitter()
     this.watching = false
+    this.projectPath = projectPath
     this.subscriptions = new CompositeDisposable()
 
     this.subscriptions.add(this.emitter)
@@ -39,21 +37,17 @@ class Motion {
     })
   }
   async exists(): Promise<boolean> {
-    return await exists(this.config.dataDirectory)
+    return await FS.exists(Path.join(this.projectPath, '.motion.json'))
   }
   async watch(terminal: boolean = false): Promise<Disposable> {
     if (!await this.exists()) {
       throw new MotionError(ERROR_CODE.NOT_MOTION_APP)
     }
 
-    this.state.get().running = true
-    this.state.get().process_id = process.pid
-    await this.state.write()
-
     if (terminal) {
       this.cli.activate()
     }
-    const pundle = await getPundleInstance(this.state, this.config, this.cli, terminal, true, error => {
+    const pundle = await getPundleInstance(this.cli, terminal, this.projectPath, true, this.config.config, error => {
       this.emitter.emit('did-error', error)
     })
     await pundle.activate()
@@ -61,8 +55,6 @@ class Motion {
       pundle.pundle.clearCache()
     })
     const disposable = new Disposable(() => {
-      this.state.get().running = false
-      this.state.write()
       this.subscriptions.remove(disposable)
       pundle.dispose()
       reloadHook.dispose()
@@ -75,20 +67,22 @@ class Motion {
     if (!await this.exists()) {
       throw new MotionError(ERROR_CODE.NOT_MOTION_APP)
     }
-    const compilation = await getPundleInstance(this.state, this.config, this.cli, terminal, false, error => {
+    const compilation = await getPundleInstance(this.cli, terminal, this.projectPath, false, this.config.config, error => {
       this.emitter.emit('did-error', error)
     })
     await compilation.compile()
-    await mkdir(Path.join(this.config.dataDirectory, '_'))
-    await writeFile(Path.join(this.config.dataDirectory, '_/bundle.js'), compilation.generate().contents)
+    await FS.mkdir(Path.join(this.config.getPublicDirectory(), '_'))
+    await FS.writeFile(Path.join(this.config.getPublicDirectory(), '_/bundle.js'), compilation.generate().contents)
   }
   async init(): Promise<void> {
     if (await this.exists()) {
       throw new MotionError(ERROR_CODE.ALREADY_MOTION_APP)
     }
-    await mkdir(this.config.rootDirectory)
-    await copy(Path.normalize(Path.join(__dirname, '..', 'template')), this.config.rootDirectory)
-    await this.state.write()
+    await FS.mkdir(this.config.getBundleDirectory())
+    await FS.mkdir(this.config.getPublicDirectory())
+    await FS.copy(Path.normalize(Path.join(__dirname, '..', 'template', 'bundle')), this.config.getBundleDirectory())
+    await FS.copy(Path.normalize(Path.join(__dirname, '..', 'template', 'public')), this.config.getPublicDirectory())
+    await this.config.write()
   }
   onDidError(callback: Function): Disposable {
     return this.emitter.on('did-error', callback)
@@ -97,13 +91,8 @@ class Motion {
     this.subscriptions.dispose()
   }
 
-  static async create(config: Motion$Config): Promise<Motion> {
-    if (await exists(config.rootDirectory)) {
-      config.rootDirectory = await realpath(config.rootDirectory)
-    }
-    fillConfig(config)
-    const state = await State.create(Path.join(config.dataDirectory, 'state.json'), Path.join(config.dataDirectory, 'config.json'))
-    return new Motion(state, config)
+  static async create(projectRoot: string): Promise<Motion> {
+    return new Motion(projectRoot, await Config.create(projectRoot))
   }
 }